Fix input lag
Previously there was a noticeable lag when drawing long lines quickly, and when drawing while fullscreened.
By operating on each event before rendering the image again, we have removed this lag.
This does have a side effect of increasing idle cpu usage. We may have to add a sleep system to compensate.